Inventory/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.5
django-crispy-forms==1.9.0
django-crudbuilder==0.2.7
django-enhanced-cbv==0.1.dev0
django-filter==2.2.0
django-multiselectfield==0.1.12
django-tables2==2.3.1
Pillow==7.1.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
wincertstore==0.2
wtf==0.1
